--- app/server/server.js
@@ -1,26 +1,0 @@
-const bodyParser = require("body-parser");
-const express = require("express");
-const dotenv = require("dotenv");
-const app = express();
-dotenv.config();
-const { Configuration, PlaidApi, PlaidEnvironments } = require("plaid");
-
-const configuration = new Configuration({
-  basePath: PlaidEnvironments[process.env.PLAID_ENV],
-  baseOptions: {
-    headers: {
-      "PLAID-CLIENT-ID": process.env.PLAID_CLIENT_ID,
-      "PLAID-SECRET": process.env.PLAID_SECRET,
-    },
-  },
-});
-
-const client = new PlaidApi(configuration);
-
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-app.get("/", (req, res) => {
-  res.send(200);
-});
-//hello agin
-app.listen(8000, () => console.log("Listening on port 8000"));
