--- app/src/App.js
@@ -1,25 +1,0 @@
-import React from "react";
-import "./App.css";
-import { BrowserRouter, Switch, Route } from "react-router-dom";
-import Dashboard from "./pages/Dashboard";
-import AuthProvider from "./context/auth.js";
-import PrivateRoute from "./components/PrivateRoute";
-import Login from "./pages/Login";
-import SignUp from "./pages/SignUp";
-// import { withRouter } from "react-router";
-
-function App() {
-  return (
-    <AuthProvider>
-      <BrowserRouter>
-        <Switch>
-        <Route exact path="/login" component={Login} />
-        <Route exact path="/signup" component={SignUp} />
-        <PrivateRoute exact path="/" component={Dashboard} />
-        </Switch>
-      </BrowserRouter>
-    </AuthProvider>
-  );
-}
-
-export default App;
