--- app/src/pages/Family.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Family() {
-    return (
-        <div className='family'>
-            <h1>family</h1>
-        </div>
-    )
-}
-
-export default Family
