--- app/src/pages/Mates.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Mates() {
-    return (
-        <div className='mates'>
-            <h1>mates</h1>
-        </div>
-    )
-}
-
-export default Mates
