+++ frontend/package.json
@@ -3,12 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@babel/eslint-parser": "^7.5.4",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "body-parser": "^1.19.1",
     "bootstrap": "^5.1.3",
     "dotenv": "^10.0.0",
+    "eslint": "^8.7.0",
     "express": "^4.17.2",
     "firebase": "^9.6.1",
     "nodemon": "^2.0.15",
@@ -28,7 +30,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "server": "nodemon ./server/server.js"
+    "server": "nodemon ./server/server.js",
+    "lint": "./node_modules/.bin/eslint ./"
   },
   "eslintConfig": {
     "extends": [
