+++ frontend/src/base.js
@@ -1,6 +1,6 @@
-import { initializeApp } from "firebase/app";
-import { getAuth } from "firebase/auth";
-import { config } from "./config/config";
+import { initializeApp } from 'firebase/app';
+import { getAuth } from 'firebase/auth';
+import { config } from './config/config';
 
 const firebaseConfig = {
   apiKey: config.REACT_APP_FIREBASE_KEY,
@@ -11,7 +11,7 @@ const firebaseConfig = {
   appId: config.REACT_APP_FIREBASE_APP_ID,
 };
 
-const app = initializeApp(firebaseConfig)
+const app = initializeApp(firebaseConfig);
 const auth = getAuth(app);
 
 export {auth};
