+++ frontend/src/components/NavBarDashboard.js
@@ -1,25 +1,25 @@
-import React, { useContext } from "react";
-import { Link, useHistory } from "react-router-dom";
-import { SidebarData } from "./SidebarData";
-import "./NavBar.css";
-import { IconContext } from "react-icons";
-import logo from "./logo.png";
-import { signOut } from "firebase/auth";
-import { auth } from "../base";
-import { AuthContext } from "../context/auth";
-import { Button } from "react-bootstrap";
-import * as ReactStrap from "reactstrap";
+import React, { useContext } from 'react';
+import { Link, useHistory } from 'react-router-dom';
+import { SidebarData } from './SidebarData';
+import './NavBar.css';
+import { IconContext } from 'react-icons';
+import logo from './logo.png';
+import { signOut } from 'firebase/auth';
+import { auth } from '../base';
+import { AuthContext } from '../context/auth';
+import { Button } from 'react-bootstrap';
+import * as ReactStrap from 'reactstrap';
 
 function NavBarDashboard() {
   const history = useHistory();
   const { user } = useContext(AuthContext);
   const handleSignout = async () => {
     await signOut(auth);
-    history.replace("/login");
+    history.replace('/login');
   };
   return (
     <>
-      <IconContext.Provider value={{ color: "#fff" }}>
+      <IconContext.Provider value={{ color: '#fff' }}>
         <div className="user-nav">
           <ReactStrap.Navbar dark expand="md" variant="light">
             <ReactStrap.NavbarBrand
@@ -39,12 +39,12 @@ function NavBarDashboard() {
             </ReactStrap.Nav>
           </ReactStrap.Navbar>
         </div>
-        <nav className={"nav-menu active"}>
+        <nav className={'nav-menu active'}>
           <ul className="nav-menu-items">
             <li >
               <Link to="#" className="navbar-toggle" >
-               
-              <img src={logo} alt="Logo" className="logo" />
+
+                <img src={logo} alt="Logo" className="logo" />
               </Link>
             </li>
             {SidebarData.map((item, index) => {
