+++ frontend/src/components/PrivateRoute.js
@@ -1,7 +1,7 @@
 
-import React, { useContext } from "react";
-import { AuthContext } from "../context/auth";
-import { Redirect, Route } from "react-router-dom";
+import React, { useContext } from 'react';
+import { AuthContext } from '../context/auth';
+import { Redirect, Route } from 'react-router-dom';
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
   const { user } = useContext(AuthContext);
