+++ frontend/src/config/config.example.js
@@ -1,0 +1,8 @@
+export const config = {
+  'REACT_APP_FIREBASE_KEY':'XXXXXXXXXXXXXXX',
+  'REACT_APP_FIREBASE_DOMAIN':'XXXXXXXXXXXXXXX',
+  'REACT_APP_FIREBASE_PROJECT_ID':'XXXXXXXXXXXXXXX',
+  'REACT_APP_FIREBASE_STORAGE_BUCKET':'XXXXXXXXXXXXXXX',
+  'REACT_APP_FIREBASE_SENDER_ID':'XXXXXXXXXXXXXXX',
+  'REACT_APP_FIREBASE_APP_ID':'XXXXXXXXXXXXXXX'
+};
