+++ frontend/src/pages/Dashboard.js
@@ -2,12 +2,12 @@ import React from 'react';
 import NavBarDashboard from '../components/NavBarDashboard';
 function Dashboard() {
 
-    return (
-        <> 
-            <NavBarDashboard/>
-            <h1>dashboard</h1>
-        </>
-    )
+  return (
+    <>
+      <NavBarDashboard/>
+      <h1>dashboard</h1>
+    </>
+  );
 }
 
 export default Dashboard;
