+++ frontend/src/pages/Login.js
@@ -1,13 +1,13 @@
-import React, { useState } from "react";
-import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../base";
-import { useHistory } from "react-router-dom";
+import React, { useState } from 'react';
+import { signInWithEmailAndPassword } from 'firebase/auth';
+import { auth } from '../base';
+import { useHistory } from 'react-router-dom';
 import './Unauthed.css';
 
 const Login = () => {
   const [data, setData] = useState({
-    email: "",
-    password: "",
+    email: '',
+    password: '',
     error: null,
     loading: false,
   });
@@ -24,18 +24,18 @@ const Login = () => {
     e.preventDefault();
     setData({ ...data, error: null, loading: true });
     if (!email || !password) {
-      setData({ ...data, error: "All fields are required" });
+      setData({ ...data, error: 'All fields are required' });
     }
     try {
       signInWithEmailAndPassword(auth, email, password);
 
       setData({
-        email: "",
-        password: "",
+        email: '',
+        password: '',
         error: null,
         loading: false,
       });
-      history.replace("/");
+      history.replace('/');
     } catch (err) {
       setData({ ...data, error: err.message, loading: false });
     }
@@ -64,43 +64,12 @@ const Login = () => {
             />
             {error ? <p className="error">{error}</p> : null}
             <button className="form-button" disabled={loading}>
-              {loading ? "Logging in ..." : "Login"}
+              {loading ? 'Logging in ...' : 'Login'}
             </button>
           </form>
         </container>
       </div>
     </div>
-    // <div className="login-signup">
-    //   <h3>log in</h3>
-    //   <form className="form" onSubmit={handleSubmit}>
-    //     <div className="input_container">
-    //       <label htmlFor="email"></label>
-    //       <input
-    //         type="text"
-    //         name="email"
-    //         value={email}
-    //         onChange={handleChange}
-    //         placeholder="email"
-    //       />
-    //     </div>
-    //     <div className="input_container">
-    //       <label htmlFor="password"></label>
-    //       <input
-    //         type="password"
-    //         name="password"
-    //         value={password}
-    //         onChange={handleChange}
-    //         placeholder="password"
-    //       />
-    //     </div>
-    //     {error ? <p className="error">{error}</p> : null}
-    //     <div className="btn_container">
-    //       <button className="btn" disabled={loading}>
-    //         {loading ? "Logging in ..." : "login"}
-    //       </button>
-    //     </div>
-    //   </form>
-    // </div>
   );
 };
