+++ frontend/src/pages/SignUp.js
@@ -1,21 +1,20 @@
-import { createUserWithEmailAndPassword } from "firebase/auth";
-import React, { useState } from "react";
-import { withRouter, useHistory} from "react-router";
-import { Link } from "react-router-dom";
-import { auth } from "../base";
+import { createUserWithEmailAndPassword } from 'firebase/auth';
+import React, { useState } from 'react';
+import { withRouter, useHistory} from 'react-router';
+import { Link } from 'react-router-dom';
+import { auth } from '../base';
 import './Unauthed.css';
 const SignUp = () => {
   const [data, setData] = useState({
-    name: "",
-    email: "",
-    password: "",
+    name: '',
+    email: '',
+    password: '',
     error: null,
     loading: false,
   });
   const history = useHistory();
   const { name, email, password, error, loading } = data;
   const handleChange = (e) => {
-    console.log(data);
     setData({ ...data, [e.target.name]: e.target.value });
   };
 
@@ -23,27 +22,27 @@ const SignUp = () => {
     // event.preventDefault();
     setData({ ...data, error: null, loading: false });
     if (!name || !email || !password) {
-      setData({ ...data, error: "All fields are required" });
+      setData({ ...data, error: 'All fields are required' });
     }
     try {
-     createUserWithEmailAndPassword(
+      createUserWithEmailAndPassword(
         auth,
         email,
         password
       );
       setData({
-        name: "",
-        email: "",
-        password: "",
+        name: '',
+        email: '',
+        password: '',
         error: null,
         loading: false,
       });
-      history.push("/");
+      history.push('/');
     } catch (err) {
       setData({ ...data, error: err.message, loading: false });
     }
-    
-    
+
+
   };
 
   return (
